Added placeholder vars/vagrant.yml
This file doesn't override any of the defaults that are
specified in each role. It's just as a default fallback
when provisioning vagrant to not load user.yml

Also if user.yml isn't defined (which it isn't for a fresh checkout)
the vagrant provision should not fail. However the deploy to production
should fail cause it's is required that we override the defaults,
like the db password and so on.